re-expose mysql form db container on 3319 (3306+13) and access dev genetracker
The offset of 13 avoids overlap with the usual mysql port if that's
running on the host. Also avoids overlap with the genetracker dev
port (which defaults to 3318).

`host.docker.internal` works by default on Docker Desktop, but not
under linux or colima, so added relevant `extra_hosts` to the compose
file.

Also added appropriate defaults to .env.example for GT_DB_* variables