put backend demographic field handling through the DRYer
Re-use list of terms in (in $demographics_private_fields if Person.php) in
fillable, hidden, and in the PersonResouce. This is error-proofing by
reducing the places that future changes might need to be made.

Noting (along with informative comment) that demographics_completed_date
is a special case of a demographics-related field that needs to be
non-private.