limit only access to address/phone information to self, admin, super-admin
As decided on 2024-07-24, timezone could be useful to have available,
and country is probably not identifying enough to worry about.

Implementing this in the most straightforward way required moving timezone
around in the display, but also fixed a css spacing issue at the same time.

Also removed some commented-out/deleted code