remove extraneous import of GroupEvent
This should not been needed since the GroupEvent is in the file's
namespace. Furthermore, it had been added to the previous commit
spuriously (not related to the other parts of the commit.

This should probably be a "fixup" to prior commit before eventual
merge to main, but keeping in history for now for tracking purposes