switch ReviewMembership to use "inject"ed group variable
This is a somewhat blunt way to try to address the reactivity issues
with display of members in GCEPs. Using vuex/pinia would be better, but
this would get things working for now.

Note that in order to do this, I needed to march up the component hierarchy
to make sure that `group` was "provide"d anywhere ReviewMembership would
eventually be used as a subcomponent.