remove runInBackground from dx consuming
This is an attempt to solve an issue where the dx:consume job keeps
holding the mutex after it should be done already. If this doesn't
resolve things, next try could be to remove the withoutOverlapping.

Will need to check to make sure this doesn't prevent other tasks from
being performed (e.g., COI reminders) if dx updates are taking too long
in a given scheduler run.