Kludge to make RichTextEditor work with both HTML and Markdown, use HTML for many legacy things
When the website team wanted us to go back to having the description texts in markdown,
the quick change I made broke the fields that used RichTextEditor for HTML (e.g., in email template
during application review). This mostly addresses this.

FIXME: there's something weird in the prop/emit handling here with the prosekit element. When
typing, it emits update for modelValue, which gets passed up then back down. I've kludged
around this for now by `watch`ing the modelValue and only updating the prosekit view if the
contents are different. This avoids a gnarly bug where the cursor would change while typing, at
the expense of hurting performance a fair bit.