artisan command to generate some example messages.
This probably doesn't change anything when being run, but just to
be on the safe side, I put a guard against it running in production.

It may be better long-term to do this by using fake data and factories,
but that would definitely need to be dev-only since it would result
in database writes.